Bundle dom-mutator into the SDK instead of requiring a 2nd import
