+++ packages/sdk-js/package.json
@@ -17,6 +17,7 @@
   "engines": {
     "node": ">=10"
   },
+  "sideEffects": false,
   "repository": {
     "type": "git",
     "url": "https://github.com/growthbook/growthbook",
