--- packages/sdk-js/src/bundle.ts
@@ -1,2 +1,0 @@
-export * from "./index";
-export { mutator } from "./mutator";
